changed line order

+++ install.sh
@@ -1,14 +1,15 @@
  #!/bin/zsh
 
 # this script will automaticly install the dotfiles from the repo and back up the old ones
+# to execute just type ./install.sh
 
 ########## Variables
-
-dir=$( cd -P -- "$(dirname -- "$(command -v -- "$0")")" && pwd -P )                    # dotfiles directory
-olddir="$HOME/dotfiles_old"             # old dotfiles backup directory
-files=('.vimrc' '.vim' '.zshrc' '.pythonrc' '.vim_python_style')    # list of files/folders to symlink in homedir
-
-echo "we are in $dir"
+# list of files/folders to symlink in homedir
+files=('.vimrc' '.vim' '.zshrc' '.pythonrc' '.vim_python_style')
+# dotfiles directory
+dir=$( cd -P -- "$(dirname -- "$(command -v -- "$0")")" && pwd -P )
+# old dotfiles backup directory
+olddir="$HOME/dotfiles_old"
 
 ##########
 
@@ -19,7 +20,9 @@ mkdir -p $olddir
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks
 echo "Moving any existing dotfiles from $HOME to $olddir"
 for file in $files; do
-    mv $HOME/$file $olddir
+    if [ -f $HOME/$file ]; then
+        mv $HOME/$file $olddir
+    fi
     echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file $HOME/$file
+    ln -s $dir/$file $HOME/
 done
